cartServer/requirements.txt
asgiref==3.7.2
Django==4.2.7
django-cors-headers==4.3.0
djangorestframework==3.14.0
dnspython==2.4.2
drf-yasg==1.21.7
gunicorn==21.2.0
ifaddr==0.2.0
inflection==0.5.1
packaging==23.2
psycopg2==2.9.9
py-eureka-client==0.11.8
pytz==2023.3.post1
PyYAML==6.0.1
sqlparse==0.4.4
tzdata==2023.3
uritemplate==4.1.1
